+++ lib/js_widget_runtime.dart
@@ -8,6 +8,7 @@ export 'src/model/js_runtime_config.dart';
 export 'src/model/widget_manifest.dart' show WidgetManifest;
 export 'src/renderer/json_widget_renderer.dart' show JsonWidgetRenderer;
 export 'src/renderer/external_asset_resolver.dart' show ExternalAssetResolver;
+export 'src/renderer/flame_asset_bundle.dart' show jsrSetFlameAssetBundle;
 export 'src/renderer/font/js_font_loader.dart' show JsFontLoader;
 export 'src/renderer/font/js_font_resolver.dart' show JsFontResolver;
 export 'src/renderer/json_widget_theme.dart' show JsonWidgetTheme;